Switch sub-sequences to a pending flag
It only matters whether a sequence is pending, not how many times we've
been waiting for it to become available. This will reduce the number of
batches produced to 1 while waiting for a subsequent sequence, still
allowing the previous (full) sequence to be flushed according to
watermark rules (specifically "0 seconds", which will otherwise wait for
the next batch to arrive).