Extract shared implicit classes
When in a trait being inherited from, the implicit classes result in
Spark serialization errors.

Since implicit classes are no longer declared within the scope of the
tile generating object, cell sizes and layout schemes need to be passed.
These are done as implicits for simplicity, although the presence of
rows and cols complicated things enough to warrant changing them to
"cells" (esp. since we're always generating square output).