Generate faceted edit histogram tiles when 'facets is present
'facets is a column containing Map[String, Int], or key -> value.
FacetedEditHistogramTileCreator creates facets for different types of
edits based on the combined set of tags present on nodes and the ways
they're associated with. Way metadata-only changes are considered "minor
versions" in this context, as the effective purpose of a node may change
even if it doesn't change itself.