Add methods to pull AOI data from postgres table
Switched to a Config object to pull all postgres
configuration from environment variables. Each
part of the connection is its own variable so that
its easier to view, configure and diff individual
parts of the connection string, learned the hard
way in other projects.

Added in the methods that get and set the sequence
value for the desired interval.

Switched Interval to a sealed trait and case object
for easier matching and type safety.

An automated scalafmt snuck its way in there in the
middle of my efforts. Untangling it was a bit of a
headache so I just left it, even though it should
be a separate commit.

Filter notification list by interval and expiration,
so that we only end up joining the unique notification +
User combinations we need to send.

Also removes the custom WeeklyInfo and DailyInfo summary
case classes in the interest of generating the simplest
report possible, using the aggregated metrics from what
was the WeeklyInfo case class.