Tweak vectortile task resources
Switch to batch size == 1 in order to
reduce memory pressure for large changesets.
The stream still processes fast enough at
this batch size.

Bump memory available to each process in
order to not crash on larger changesets.
This is a bit of an unfounded change in that
we have no proof that these processes won't
still crash on massive changesets.